Split CosmosDBShellExtension.ts into focused modules
Extract subsystems from the 1295-line CosmosDBShellExtension.ts into

sibling modules so each file has a single responsibility:

- shellCommand.ts: command resolution, path validation, watchForEarlyExit, quoteArg, COSMOS_DB_SHELL_TERMINAL_NAME

- shellSupportCache.ts: cached '--version' probing (isCosmosDBShellInstalled, getDetectedCosmosDBShellVersion, invalidateCosmosDBShellSupportCache)

- nodeCredentials.ts: AuthKind, credential accessors, getNodeAuthKind, getCosmosDBShellToken

- terminalReuse.ts: ShellTerminalState, terminalStates map, canReuse/findReusable, buildInteractiveConnectCommand

- install/dotNetSdk.ts: SDK constants, compareDotNetVersions, hasRequiredDotNetSdk, tryInstallDotNetSdkViaExtension

- install/installPrompts.ts: install/repair prompts; takes a launchShell callback to avoid a circular import

- mcpProvider.ts: registerMcpServer, port probing, resolveMcpServer, config-change wiring

- languageServer.ts: registerCosmosDBShellLanguageServer

CosmosDBShellExtension.ts (314 lines) keeps the extension class, launchCosmosDBShell, connectCosmosDBShell, and re-exports the public surface consumed by extension.ts.

Carried in passing: void-prefixed two fire-and-forget setContext calls; replaced repeated 'Cosmos DB Shell' literal with COSMOS_DB_SHELL_TERMINAL_NAME; replaced repeated 6128 with DEFAULT_MCP_PORT; commented the useMcp = mcpEnabled && !foundTerminal rationale; removed a redundant duplicate authKind assignment in connectCosmosDBShell.

No behavior, telemetry, or user-facing string changes.
